requirements.txt
python-telegram-bot[job-queue]==21.6
Pillow==10.4.0
supabase==2.11.0
python-dotenv==1.0.1
aiohttp==3.11.11
requests==2.32.3
